# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API per il portale di orientamento universitario di Sofia
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [public]
      summary: Elenco dei corsi di laurea
      parameters:
        - name: university
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Corsi disponibili
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
  /orientation-requests:
    post:
      operationId: createOrientationRequest
      tags: [public]
      summary: Invia una richiesta di orientamento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrientationRequestInput"
      responses:
        "201":
          description: Richiesta registrata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrientationRequestReceipt"
  /tour-slots:
    get:
      operationId: listTourSlots
      tags: [public]
      summary: Elenca gli orari disponibili per una data
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Fasce orarie, con disponibilità
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TourSlot"
  /appointment-slots:
    get:
      operationId: listAppointmentSlots
      tags: [public]
      summary: Elenca gli orari liberi per consulenze e tour
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Slot orari con disponibilità condivisa nel calendario di Sofia
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppointmentSlot"
  /tour-bookings:
    post:
      operationId: createTourBooking
      tags: [public]
      summary: Prenota un tour della piattaforma
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TourBookingInput"
      responses:
        "201":
          description: Prenotazione registrata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TourBookingReceipt"
...
components:
  schemas:
...
    Course:
      type: object
      properties:
        id: { type: string }
        university: { type: string }
        category: { type: string }
        name: { type: string }
        duration: { type: string }
      required: [id, university, category, name, duration]
    ContactInput:
      type: object
      properties:
        firstName: { type: string, minLength: 1, maxLength: 80 }
        lastName: { type: string, minLength: 1, maxLength: 80 }
        email: { type: string, format: email, maxLength: 254 }
        province: { type: string, minLength: 2, maxLength: 60 }
        phone: { type: string, minLength: 6, maxLength: 30 }
      required: [firstName, lastName, email, province, phone]
    OrientationRequestInput:
      allOf:
        - $ref: "#/components/schemas/ContactInput"
        - type: object
          properties:
            university: { type: string, minLength: 1 }
            courseId: { type: string, minLength: 1 }
            date: { type: string, format: date }
            time: { type: string, pattern: "^[0-2][0-9]:[0-5][0-9]$" }
            captchaToken: { type: string, minLength: 1, maxLength: 4096 }
          required: [university, courseId, date, time, captchaToken]
...
    OrientationRequestReceipt:
      type: object
      properties:
        id: { type: integer }
        date: { type: string, format: date }
        time: { type: string }
        meetUrl: { type: string, format: uri }
        createdAt: { type: string, format: date-time }
        confirmationEmailStatus: { $ref: "#/components/schemas/ConfirmationEmailStatus" }
      required: [id, date, time, meetUrl, createdAt, confirmationEmailStatus]
    TourBookingInput:
      allOf:
        - $ref: "#/components/schemas/ContactInput"
        - type: object
          properties:
            date: { type: string, format: date }
            time: { type: string, pattern: "^[0-2][0-9]:[0-5][0-9]$" }
            captchaToken: { type: string, minLength: 1, maxLength: 4096 }
          required: [date, time, captchaToken]
...
    TourBookingReceipt:
      type: object
      properties:
        id: { type: integer }
        date: { type: string, format: date }
        time: { type: string }
        meetUrl: { type: string, format: uri }
        confirmationEmailStatus: { $ref: "#/components/schemas/ConfirmationEmailStatus" }
      required: [id, date, time, meetUrl, confirmationEmailStatus]
    TourSlot:
      type: object
      properties:
        time: { type: string }
        available: { type: boolean }
      required: [time, available]
    AppointmentSlot:
      type: object
      properties:
        time: { type: string }
        available: { type: boolean }
      required: [time, available]
...
    ConfirmationEmailStatus:
      type: string
      enum: [sent, failed, not_configured, disabled]
...