# lib/api-zod/src/generated/types/orientationRequestInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API per il portale di orientamento universitario di Sofia
 * OpenAPI spec version: 0.1.0
 */
import type { ContactInput } from './contactInput';

export type OrientationRequestInput = ContactInput & {
  /** @minLength 1 */
  university: string;
  /** @minLength 1 */
  courseId: string;
  date: Date;
  /** @pattern ^[0-2][0-9]:[0-5][0-9]$ */
  time: string;
  /**
     * @minLength 1
     * @maxLength 4096
     */
  captchaToken: string;
};